+++ pages/admin/annotation/examples/index.vue
@@ -22,7 +22,7 @@
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator';
 import Example from '~/components/Example.vue';
-import { NewExample } from '~/assets/util';
+import { NewExample } from '~/plugins/util';
 import { Auth } from 'aws-amplify';
 
 @Component({
