+++ pages/admin/annotation/tweets/index.vue
@@ -23,7 +23,7 @@
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator';
 import Example from '~/models/Example'
-import { NewExample } from '~/assets/util';
+import { NewExample } from '~/plugins/util';
 import { Auth } from 'aws-amplify';
 
 @Component({
